pyproject.toml
[project]
name = "agent-reach"
version = "1.13.0"
description = "Windows-first Twitter/X research capability layer with a Python SDK and JSON CLI."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "Neo Reid"}]
keywords = [
    "codex", "windows", "research", "integration",
    "python-sdk", "json-cli", "github-actions", "discord-bot",
    "twitter", "x.com",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "requests>=2.28",
    "feedparser>=6.0",
    "browser-cookie3>=0.20",
    "loguru>=0.7",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
    "mypy>=1.12",
    "types-requests>=2.32",
    "types-PyYAML>=6.0",
]

[project.scripts]
agent-reach = "agent_reach.cli:main"

[project.urls]
Upstream = "https://github.com/Panniantong/Agent-Reach"
Repository = "https://github.com/iwachacha/twitter-reach"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent_reach"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
ignore_missing_imports = true
disable_error_code = ["import-untyped"]
exclude = ["^tests/"]
